Move all 'views' into their own directory
